static-analyzer/pyproject.toml
[project]
name = "zorbox-static-analyzer"
version = "0.1.0"
description = "ZORBOX Static Analyzer (MVP)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "prometheus-client>=0.20.0",
  "python-multipart>=0.0.9",
  "pefile>=2023.2.7",
  "yara-python>=4.3.1",
  "oletools>=0.60.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
